Add files via upload
1.3.0 includes the following updates for the following issues 
Handling new operators in 2016 UR1
404 Error when activating Routing Rule Sets.
Unable to Login to CRM for Phone / Tablet - Disabled IQ for the following entities 
 Channel Access Profile Rule Item - 9401
 External Party Item - 9987
 Sync Error - 9869
Remove metadata caching - Recommended by AppSource. Metadata caching is on-demand by entity
Charting using CRMIQ views - This release includes saving Saved Views with the updated FetchXml based on the query specified. This means that saving queries with the "Not In" condition will result in queries that are no longer editable but will be usable with Charts. A future release will include an option to disable updating the underlying query when saving a user query. However, meeting this requirement in the short term adds a lot of value to IQ for certain users so we're releasing this ahead of the configurable option.